requirements.txt
python-dotenv
requests
fastapi
uvicorn
transformers
torch
diffusers
scipy
serpapi
python-dotenv
pillow
openai
numpy==1.26.1
nltk